[lib-chat] Refactor error handling into separate module
- Extracted `ErrorKind`, `Error`, and `Result` types into a new `error.rs` module
- Moved previous error implementation from `lib.rs` to `error.rs`
- Added `pub use` statements in `lib.rs` to re-export the error types
